refactor: extract CreateTaskInputData, CreateTaskOutputBoundary, CreateTaskOutputData, MarkTaskCompleteInputData, MarkTaskCompleteOutputBoundary, MarkTaskCompleteOutputData to separate file
