+++ src/main/java/use_case/CreateTaskOutputBoundary.java
@@ -1,0 +1,9 @@
+package use_case;
+
+/**
+ * Output boundary interface for task creation.
+ */
+public interface CreateTaskOutputBoundary {
+    void presentSuccess(CreateTaskOutputData outputData);
+    void presentError(String error);
+}
