+++ src/main/java/use_case/CreateTaskUseCase.java
@@ -62,60 +62,3 @@ public class CreateTaskUseCase {
     }
 }
 
-/**
- * Input data for creating a task.
- */
-class CreateTaskInputData {
-    private final String name;
-    private final String description;
-    private final String category;
-    private final Task.Priority priority;
-    private final LocalDate beginDate;
-    private final LocalDate dueDate;
-
-    public CreateTaskInputData(String name, String description, String category,
-                               Task.Priority priority, LocalDate beginDate, LocalDate dueDate) {
-        this.name = name;
-        this.description = description;
-        this.category = category;
-        this.priority = priority;
-        this.beginDate = beginDate;
-        this.dueDate = dueDate;
-    }
-
-    // Getters
-    public String getName() { return name; }
-    public String getDescription() { return description; }
-    public String getCategory() { return category; }
-    public Task.Priority getPriority() { return priority; }
-    public LocalDate getBeginDate() { return beginDate; }
-    public LocalDate getDueDate() { return dueDate; }
-}
-
-/**
- * Output data for task creation.
- */
-class CreateTaskOutputData {
-    private final String taskId;
-    private final String taskName;
-    private final String message;
-
-    public CreateTaskOutputData(String taskId, String taskName, String message) {
-        this.taskId = taskId;
-        this.taskName = taskName;
-        this.message = message;
-    }
-
-    // Getters
-    public String getTaskId() { return taskId; }
-    public String getTaskName() { return taskName; }
-    public String getMessage() { return message; }
-}
-
-/**
- * Output boundary interface for task creation.
- */
-interface CreateTaskOutputBoundary {
-    void presentSuccess(CreateTaskOutputData outputData);
-    void presentError(String error);
-}
