+++ src/main/java/use_case/MarkTaskCompleteOutputBoundary.java
@@ -1,0 +1,9 @@
+package use_case;
+
+/**
+ * Output boundary interface for marking task complete.
+ */
+public interface MarkTaskCompleteOutputBoundary {
+    void presentSuccess(MarkTaskCompleteOutputData outputData);
+    void presentError(String error);
+}
