+++ src/main/java/use_case/MarkTaskCompleteUseCase.java
@@ -71,47 +71,3 @@ public class MarkTaskCompleteUseCase {
     }
 }
 
-/**
- * Input data for marking task complete.
- */
-class MarkTaskCompleteInputData {
-    private final String taskId;
-
-    public MarkTaskCompleteInputData(String taskId) {
-        this.taskId = taskId;
-    }
-
-    public String getTaskId() { return taskId; }
-}
-
-/**
- * Output data for marking task complete.
- */
-class MarkTaskCompleteOutputData {
-    private final String taskId;
-    private final String taskName;
-    private final LocalDateTime completionTime;
-    private final double completionRate;
-
-    public MarkTaskCompleteOutputData(String taskId, String taskName,
-                                      LocalDateTime completionTime, double completionRate) {
-        this.taskId = taskId;
-        this.taskName = taskName;
-        this.completionTime = completionTime;
-        this.completionRate = completionRate;
-    }
-
-    // Getters
-    public String getTaskId() { return taskId; }
-    public String getTaskName() { return taskName; }
-    public LocalDateTime getCompletionTime() { return completionTime; }
-    public double getCompletionRate() { return completionRate; }
-}
-
-/**
- * Output boundary interface for marking task complete.
- */
-interface MarkTaskCompleteOutputBoundary {
-    void presentSuccess(MarkTaskCompleteOutputData outputData);
-    void presentError(String error);
-}
